hiera: Specify a value for the ntp_servers hiera key
The ntp_servers key is referred to in
`modules/rhizo_base/manifests/init.pp` and was otherwise undefined. It
is in the common file since it seems similar to other parameters
already in the file, but there is no strict reason for the current
location. The assigned value is also just an initial default guess.