rhizo_base/users: Remove requirement of specific UID
When deploying on a virtualized platform, the initial login user 1000
may be automatically provisioned for SSH by the platform (this is true
with Vagrant on Arch). It seems there is no specific reason for the
rhizomatica user to be UID 1000, and code which depends on a specific
UID value instead of a user name or group name would be brittle.