rhizo_base: Specify users when using source_permissions
The source_permissions => use parameter will capture both the file
permissions from the source as well as the source file owner and group
unless owner and group are explicitly given. This allows for the
source to be mounted by an external user, and fixes an issue where
scripts in /home/rhizomatica/bin were owned by root (since according
to the readme the original puppet repo clone should be done by root).