rhizo_base/apt: Remove inconsistent notify_update
The apt::source resource defaults the parameter notify_update to
true. All other resources in the module do not explicitly specify
notify_update, so it stands out that it is set to its default
here. This commit removes the explicit setting to make the freeswith
repository consistent with the others and not unnecessarily attract
the attention of a future reader inexperienced with puppet.