rhizo_base/freeswitch: Install service after core and modules
The freeswitch install is relatively complex, with lots of
configuration files and separately packaged and dynamically loaded
modules. This commit groups the freeswitch install process into a
class, and separates out the creation and starting of service files
until after that class completes fully. This allows for puppet to
automatically start freeswitch and appropriately enable the freeswitch
service file so freeswitch will start automatically on reboot as well.