rhizo_base/init: Ensure required nitb is enabled before restart
The smpp service is managed through runit instead of systemd, and
doesn't define a soft runtime depencency on the nitb. It will fail to
start with a timeout if the nitb is not available. In the long run it
should probably be migrated to systemd so the runtime dependency can
be captured more reliably there, but for now ensure that the nitb is
running before puppet tries to automatically restart it. This prevents
a race condition where the restart would sometimes fail during
install.