rhizo_base: Add contains dependencies
Puppet treats included classes specially, and does not assume that
they are contained in their outside class for the sake of dependency
ordering. This means that any other resources requiring the base
non-distribution specific classes were actually not capturing the
ordering dependencies on resources defined in the distribution
specific class!