Improve NITB config template.
Introduces a multiple file template and iterates over the defined
BTS array from the hiera data.

This is really a backport of the as yet uncommitted work on creating osmo-*
config files for the split stack. Making this change here for the nitb will
allow for removal for some deprecated variables in hiera.

At this time, I don't think there is any site that is actually using
a puppet-generated nitb config file, but if there were, the results of this code
should be more less funcionally equivalent, providing the BTS definition is correct.